+++ server/helpers/default.js
@@ -1,0 +1,9 @@
+const {isset} = require("./isset");
+
+const defaultValue = (val, def) => {
+    return isset(val, false) ? val : def
+}
+
+module.exports = {
+    defaultValue
+}